Create README.md
Music is a branch of art that expresses emotions. It is possible to reveal the relations between the sounds of music and some emotions with certain principles. Music emotion recognition systems is used actively for various purposes including, management of personal music collections and development of music suggestion systems, in music therapy and in the treatment of emotional disorders. Determining the emotional contents of music with calculations is an interdisciplinary study, which includes not only signal processing and machine learning but also auditory perception concepts, psychology, cognitive science and musicology. Although there are a great number of studies in this field, emotion recognition from music signals is a difficult work.

One of the reasons for this difficulty is that emotion perception is subjective in essence, and people may perceive different emotions for the same song. This subjectivity problem makes the performance evaluation of a music emotion recognition system become basically difficult. Another reason is that it is not easy to define emotions in a universal manner, because the attributes that are used to define the same emotion may vary among people. Finally, it is still unknown how music arouses emotions in humans, and it has not been understood fully how the inner element of music create emotional reactions on the listener. Although there is no standardized method for recognizing and analyzing emotions in music, the psychological emotional method is used widely in many studies in this field. Current automated systems that target music emotion recognition may be grouped in two main groups categorically or dimensionally. The categorical models, which are also known as discrete models, use single words or phrases to classify emotions (e.g. happy, sad, angry, relax). In dimensional models, different emotional states are positioned in an area that is transformed into emotional states that are placed in a two-dimensional space. Thayer's two-dimensional emotion model is an example for dimensional models. This model provides a simple and highly effective model following the positioning of emotions in a two-dimensional space. The amount of stimulation and valence is measured in the model across vertical and horizontal axes, respectively. The music is categorized emotionally into a number of classes over these models that are used commonly, and following this step, a standard pattern recognition procedure is applied to train a classifier.

In the typical traditional approaches to music emotion recognition, those features representing the acoustic content of music like rhythm, tone and harmony are extracted. Different machine learning algorithms are applied for understanding the relationship between the extracted features of music and the pre-determined emotion labels. The disadvantages of traditional approaches are the difficulty of feature extraction from music recordings and uncertainty of whether extracted features are related to musical emotions. Instead of extracting acoustic features as in the traditional music emotion recognition approaches, this study offers a new perspective related music emotion recognition by deep visual features in the chroma spectrogram images.

The main contributions of this data set :

*A new dataset consisting of four classes is created.
*It was shown that the emotional content of music can be analyzed with the spectrograms obtained from music recordings.
*Chroma spectrograms showing the energy distribution around each note in music are extracted from deep features and are used for music emotion recognition.

There are certain restrictions we must consider in the development of music emotion recognition systems. Namely;

The model data sets (e.g. categorical model or dimensional model) on which the developed music emotion recognition system will be applied should be decided.

For musical perception is subjective and cannot be defined universally, a limited number of emotion labels should be used. Synonymous adjectives shouldn't be used as the emotion labels of separate classes.

Different pieces of music in the data set should be in the same format and also have the same high-quality and equal sample frequency.

Any features extracted from music data should be related with music emotions.

Pieces of music in the data set shouldn't be very short since emotions in a given piece of music become meaningful only over a certain duration.